--- src/Network/Network/FirewallPolicy/ManagedRules/Exclusion/ExclusionManagedRuleSet/ExclusionManagedRuleGroup/ExclusionManagedRule/AzureApplicationGatewayFirewallPolicyExclusionManagedRule.cs
@@ -1,41 +1,0 @@
-﻿// ----------------------------------------------------------------------------------
-//
-// Copyright Microsoft Corporation
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// http://www.apache.org/licenses/LICENSE-2.0
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// ----------------------------------------------------------------------------------
-
-using Microsoft.Azure.Commands.Network.Models;
-using System.Management.Automation;
-
-namespace Microsoft.Azure.Commands.Network
-{
-    public class AzureApplicationGatewayFirewallPolicyExclusionManagedRule : NetworkBaseCmdlet
-    {
-        [Parameter(
-            Mandatory = true,
-            HelpMessage = "Rule Id.")]
-        [ValidateNotNullOrEmpty]
-        public string RuleId { get; set; }
-
-        public override void ExecuteCmdlet()
-        {
-            base.ExecuteCmdlet();
-        }
-
-        protected PSApplicationGatewayFirewallPolicyExclusionManagedRule NewObject()
-        {
-            return new PSApplicationGatewayFirewallPolicyExclusionManagedRule()
-            {
-                RuleId = this.RuleId
-            };
-        }
-    }
-}
